Update to work with python 3

+++ script.py
@@ -1,6 +1,7 @@
 # Licensed under GPLv3, see LICENSE.txt for details
 import re
 import subprocess
+import sys
 import xbmc
 import xbmcgui
 import xbmcplugin
@@ -10,7 +11,7 @@ lockdir = 'special://temp/kodi.script.togglepower.inprogress/'
 
 exists = xbmcvfs.exists(lockdir)
 if exists:
-	xbmc.log('Kodi.Script.ToggleLGTVPower: Not toggling power; lockfile exists', xbmc.LOGNOTICE)
+	xbmc.log('Kodi.Script.ToggleLGTVPower: Not toggling power; lockfile exists', xbmc.LOGINFO)
 	# we'll remove the lockdir here, just to avoid the whole functionality seizing up
 	# if the dir is created and not destroyed for some reason. it's unlikely that a user
 	# will try to power off more than once after initiating power off, and it's better
@@ -19,7 +20,7 @@ if exists:
 	success = xbmcvfs.rmdir(lockdir)
 	if not success:
 		xbmc.log('Kodi.Script.ToggleLGTVPower: Failed to remove lock directory', xbmc.LOGERROR)
-	quit()
+	sys.exit(0)
 
 success = xbmcvfs.mkdir(lockdir)
 if not success:
@@ -28,22 +29,25 @@ if not success:
 d = xbmcgui.DialogProgressBG()
 d.create('Power', 'Just a moment while I check whether I need to turn the TV on or off')
 
-xbmc.log("Toggling TV power", xbmc.LOGNOTICE)
+xbmc.log("Toggling TV power", xbmc.LOGINFO)
 
 # The cec-client command can sometimes cause TV to wake from standby, however
 # this is fine as we only use it when it should wake if in standby
-try:
-	s = subprocess.check_output('echo pow 0 | cec-client -s -d 1 -m', shell=True)
-except:
+s = subprocess.run('echo pow 0 | ' + bin + ' -s -d 1 -m', shell=True, capture_output=True)
+if s.returncode != 0:
 	xbmc.log("Kodi.Script.ToggleLGTVPower: cec-client command to get TV status failed, may wake TV anyway", xbmc.LOGERROR)
-	d.update(100, 'Failed to get TV status, may wake TV anyway')
+	xbmc.log(s.stdout.decode(), xbmc.LOGERROR)
+	xbmc.log(s.stderr.decode(), xbmc.LOGERROR)
+	d.update(100, 'Power', 'Failed to get TV status, may wake TV anyway')
 	d.close()
+	del d
 	success = xbmcvfs.rmdir(lockdir)
 	if not success:
 		xbmc.log('Kodi.Script.ToggleLGTVPower: Failed to remove lock directory', xbmc.LOGERROR)
-	quit()
+	sys.exit(0)
+
 reStatus = re.compile(b'power status: (.+)')
-match = re.search(reStatus, s)
+match = re.search(reStatus, s.stdout)
 
 if match:
 	state = match.group(1)
@@ -51,22 +55,22 @@ else:
 	state = 'unknown'
 
 if state == 'on':
-	d.update(50, 'Turning off TV')
-	xbmc.log("Kodi.Script.ToggleLGTVPower: Turning off TV with Kodi.Script.TurnOffLGTV", xbmc.LOGNOTICE)
+	d.update(50, 'Power', 'Turning off TV')
+	xbmc.log("Kodi.Script.ToggleLGTVPower: Turning off TV with Kodi.Script.TurnOffLGTV", xbmc.LOGINFO)
 	xbmc.executebuiltin('RunAddon(kodi.script.turnofflgtv)', True)
 else:
-	d.update(50, 'Turning on TV')
-	xbmc.log("Kodi.Script.ToggleLGTVPower: Turning on TV with cec-client", xbmc.LOGNOTICE)
+	d.update(50, 'Power', 'Turning on TV')
+	xbmc.log("Kodi.Script.ToggleLGTVPower: Turning on TV with cec-client", xbmc.LOGINFO)
 	# Running cec-client has a habit of disabling Kodi's CEC connection,
 	# so we can't use its inbuilt CECActivateSource command
 	#xbmc.executebuiltin('CECActivateSource')
-	try:
-		subprocess.call('echo on 0 | cec-client -s -d 1 -m', shell=True)
-	except:
+	s = subprocess.run('echo on 0 | ' + bin + ' -s -d 1 -m', shell=True)
+	if s.returncode != 0:
 		xbmc.log("Kodi.Script.ToggleLGTVPower: cec-client command to turn on TV failed", xbmc.LOGERROR)
 
 d.update(100, 'Done')
 d.close()
+del d
 
 success = xbmcvfs.rmdir(lockdir)
 if not success:
